+++ source/psim/p_trace_sight.cpp
@@ -311,7 +311,7 @@ static bool SightCheckLine (SightTraceInfo &trace, line_t *ld) {
   const float dot2 = ld->PointDistance(trace.End);
 
   // if starting point is on a line, ignore this line
-  if (fabs(dot1) <= 0.1f) return true;
+  if (fabsf(dot1) <= 0.1f) return true;
 
   // do not use multiplication to check: zero speedup, lost accuracy
   //if (dot1*dot2 >= 0) return true; // line isn't crossed
@@ -440,7 +440,7 @@ static bool SightPathTraverse (SightTraceInfo &trace) {
   trace.Delta = trace.End-trace.Start;
   trace.Hit1S = false;
 
-  if (fabs(trace.Delta.x) <= 1.0f && fabs(trace.Delta.y) <= 1.0f) {
+  if (fabsf(trace.Delta.x) <= 1.0f && fabsf(trace.Delta.y) <= 1.0f) {
     // vertical trace; check starting sector planes and get out
     trace.Delta.x = trace.Delta.y = 0; // to simplify further checks
     trace.LineEnd = trace.End;
@@ -482,7 +482,7 @@ static bool SightPathTraverse (SightTraceInfo &trace) {
 static bool SightPathTraverse2 (SightTraceInfo &trace) {
   trace.Delta = trace.End-trace.Start;
   trace.LineStart = trace.Start;
-  if (fabs(trace.Delta.x) <= 1.0f && fabs(trace.Delta.y) <= 1.0f) {
+  if (fabsf(trace.Delta.x) <= 1.0f && fabsf(trace.Delta.y) <= 1.0f) {
     // vertical trace; check starting sector planes and get out
     trace.Delta.x = trace.Delta.y = 0; // to simplify further checks
     trace.LineEnd = trace.End;
